+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - process_changes:  Process dirty and untracked files.
 
 ### Changed
+- post_process:  Added Log class instance to function to record entries to log.
+- Added Log class instance for all post_process calls.
 - process_changes:  Replaced options for process_dirty and process_untracked to cfg settings.
 - merge:  Replaced dirty and untrack code with call to process_changes function.
 - quarantine:  Moved get_dirty and get_untracked to process_changes function.